fix(material): D-414 出库数量支持小数 + 采购明细金额按到货算、补最新到货日期
1) 出库小数链路（承接 D-410/D-413 的遗留）
- 迁移 V202709150200：t_material_outbound_log.quantity INT → DECIMAL(12,4)
- MaterialOutboundLog.quantity 改 BigDecimal
- scan-outbound 不再 intValue()，与扫码入库一致走 parseQuantity
- freeOutbound / manualOutbound 全程 BigDecimal：校验、库存比较、扣减、
  出库单金额与库存流水均按小数计算（此前 1.32 米会被截成 1，库存少扣）
- 出入库流水 DTO、仓库看板最近操作 DTO 数量改 BigDecimal（不再取整显示）
- 前端出库/领取/扫码数量输入框补齐 precision=2 / step=0.01

2) 订单物料采购明细（用户反馈）
- 金额列：由「采购数量 × 单价」改为「实际到货数量 × 单价」
  （到货 255.5 米却按 255 米算的问题），未到货行显示 "-"
- 最新到货日期：由 expectedArrivalDate（预计）改为 actualArrivalDate（真实）
- 后端每次到货都回写 actualArrivalDate（此前仅 completed 才写，
  部分到货的行该字段永远为空）
- 迁移 V202709150210：回填历史已到货但缺日期的行（优先取入库单时间）

自检：mvn compile/test-compile 通过、tsc --noEmit 通过、
eslint 全量 0 error、npm run build 通过；未触碰任何打印相关文件。

+++ backend/src/main/java/com/fashion/supplychain/intelligence/orchestration/MonthlyBizSummaryOrchestrator.java
@@ -154,7 +154,7 @@ public class MonthlyBizSummaryOrchestrator {
         oqw.ge("outbound_time", start).lt("outbound_time", end).eq("delete_flag", 0).select("quantity");
         List<MaterialOutboundLog> outbounds = materialOutboundLogMapper.selectList(oqw);
         long outboundQty = outbounds.stream()
-            .mapToLong(o -> o.getQuantity() == null ? 0 : o.getQuantity()).sum();
+            .mapToLong(o -> o.getQuantity() == null ? 0L : o.getQuantity().longValue()).sum();
 
         Map<String, Object> m = new LinkedHashMap<>();
         m.put("inboundCount", inbounds.size());
